wip: updated modals
