+++ src/assets/icons/CopyIcon.tsx
