+++ src/assets/icons/PillCloseIcon.tsx
