+++ src/components/security/seedphrase/ViewSeedPhrase.tsx
@@ -1,7 +1,5 @@
-import { useState } from 'react';
-import { StagedSecuritySectionType } from '../../StagedSecurity';
-import { BlackButton, Text, Title } from '../../styles';
 import {
+    Wrapper,
     ButtonWrapper,
     Check,
     Checkbox,
@@ -15,16 +13,19 @@ import {
     Word,
     WordColumn,
     WordList,
-    Wrapper,
-} from './styles';
-import CopyIcon from '../../icons/CopyIcon';
-import CheckIcon from '../../icons/CheckIcon';
-import { AnimatePresence } from 'motion/react';
+    BlackButton,
+    Title,
+    Text,
+} from './styles.ts';
 import { useTranslation } from 'react-i18next';
 import { useCopyToClipboard } from '@app/hooks';
+import { useState } from 'react';
+import { AnimatePresence } from 'motion/react';
+import { useStagedSecurityStore } from '@app/store';
+import CheckIcon from '@app/assets/icons/CheckIcon.tsx';
+import CopyIcon from '@app/assets/icons/CopyIcon.tsx';
 
-interface Props {
-    setSection: (section: StagedSecuritySectionType) => void;
+interface ViewSeedPhraseProps {
     words: string[];
 }
 
@@ -36,18 +37,19 @@ const seedWordGroups = (words: string[]) => {
     return groups;
 };
 
-const SeedPhrase = ({ setSection, words }: Props) => {
+export function ViewSeedPhrase({ words }: ViewSeedPhraseProps) {
+    const setModalStep = useStagedSecurityStore((s) => s.setModalStep);
     const { t } = useTranslation('staged-security');
     const { isCopied, copyToClipboard } = useCopyToClipboard();
     const [checked, setChecked] = useState(false);
 
-    const handleCopy = () => {
+    function handleCopy() {
         copyToClipboard(words.join(' '));
-    };
+    }
 
-    const handleCheckboxClick = () => {
-        setChecked(!checked);
-    };
+    function handleCheckboxClick() {
+        setChecked((c) => !c);
+    }
 
     return (
         <Wrapper>
@@ -102,12 +104,10 @@ const SeedPhrase = ({ setSection, words }: Props) => {
                     <CheckboxText>{t('seedPhrase.checkbox')}</CheckboxText>
                 </CheckboxWrapper>
 
-                <BlackButton onClick={() => setSection('VerifySeedPhrase')} disabled={!checked}>
+                <BlackButton onClick={() => setModalStep('VerifySeedPhrase')} disabled={!checked}>
                     <span>{t('seedPhrase.button')}</span>
                 </BlackButton>
             </ButtonWrapper>
         </Wrapper>
     );
-};
-
-export default SeedPhrase;
+}
