+++ src/containers/floating/FloatingElements.tsx
@@ -1,7 +1,6 @@
 import { FloatingTree } from '@floating-ui/react';
 
 import SettingsModal from './Settings/SettingsModal.tsx';
-// import StagedSecurity from './StagedSecurity/StagedSecurity.tsx';
 import AutoUpdateDialog from './AutoUpdateDialog/AutoUpdateDialog.tsx';
 import ExternalDependenciesDialog from './ExternalDependenciesDialog/ExternalDependenciesDialog.tsx';
 import CriticalErrorDialog from './CriticalErrorDialog/CriticalErrorDialog.tsx';
@@ -24,6 +23,7 @@ import KeychainDialog from './Keychain/KeychainDialog.tsx';
 import PinDialog from '@app/containers/floating/security/pin/PinDialog.tsx';
 import ForgotPinDialog from './security/pin/ForgotPinDialog.tsx';
 import SecurityPromptDialog from '@app/containers/floating/security/prompt/SecurityPromptDialog.tsx';
+import SeedPhrase from '@app/containers/floating/security/seedphrase/SeedPhrase.tsx';
 
 const environment = import.meta.env.MODE;
 
@@ -32,7 +32,6 @@ const FloatingElements = () => {
         <FloatingTree>
             {environment === 'development' && <AdminUI />}
             <SettingsModal />
-            {/*<StagedSecurity />*/}
             <AutoUpdateDialog />
             <CriticalErrorDialog />
             <ExternalDependenciesDialog />
@@ -53,6 +52,7 @@ const FloatingElements = () => {
             <PinDialog />
             <ForgotPinDialog />
             <SecurityPromptDialog />
+            <SeedPhrase />
         </FloatingTree>
     );
 };
