+++ src/containers/floating/security/common.styles.ts
