+++ src/containers/floating/security/pin/PinDialog.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { emit } from '@tauri-apps/api/event';
 
-import { setDialogToShow, useUIStore } from '@app/store';
+import { setDialogToShow, useStagedSecurityStore, useUIStore } from '@app/store';
 import { Dialog, DialogContent } from '@app/components/elements/dialog/Dialog.tsx';
 import CloseButton from '@app/components/elements/buttons/CloseButton.tsx';
 
@@ -15,7 +15,10 @@ export default function PinDialog() {
     const { t } = useTranslation('wallet');
     const [showComplete, setShowComplete] = useState(false);
     const dialogToShow = useUIStore((s) => s.dialogToShow);
-    const isOpen = dialogToShow === 'enterPin' || dialogToShow === 'createPin';
+    const step = useStagedSecurityStore((s) => s.step);
+    const showModal = useStagedSecurityStore((s) => s.showModal);
+    const createOpen = showModal && step === 'CreatePin';
+    const isOpen = dialogToShow === 'enterPin' || createOpen;
 
     function handleClose() {
         void emit('pin-dialog-response', { pin: undefined });
